# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenClaw Pairing and Communication API
  version: 0.1.0
  description: In-memory scaffold API for device register, pairing, and signaling relay.
servers:
  - url: https://api.openclawapp.dev
paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthz
      responses:
        "200":
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /v1/devices/register:
    post:
      summary: Register desktop device basic info
      operationId: registerDevice
      responses:
        "200":
          description: Registered
  /v1/devices/heartbeat:
    post:
      summary: Update desktop online heartbeat
      operationId: heartbeatDevice
      responses:
        "200":
          description: Heartbeat accepted
  /v1/devices/{deviceId}/status:
    get:
      summary: Get desktop device status
      operationId: getDeviceStatus
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device status
        "404":
          description: Device not found
  /v1/pair/sessions:
    post:
      summary: Create one-time pairing session from desktop
      operationId: createPairSession
      responses:
        "200":
          description: Pair session created
  /v1/pair/claim:
    post:
      summary: Claim pairing by QR token
      operationId: claimPair
      responses:
        "200":
          description: Pair claim success
        "410":
          description: Pair session expired
  /v1/pair/claim-by-code:
    post:
      summary: Claim pairing by manual code
      operationId: claimPairByCode
      responses:
        "200":
          description: Pair claim success
        "410":
          description: Pair session expired
  /v1/pair/revoke:
    post:
      summary: Revoke a paired binding
      operationId: revokePair
      responses:
        "200":
          description: Binding revoked
        "404":
          description: Binding not found
  /v1/pair/bindings:
    get:
      summary: List user-device bindings
      operationId: listBindings
      responses:
        "200":
          description: OK
  /v1/signal/send:
    post:
      summary: Send signaling or relay event to target peer
      operationId: sendSignal
      responses:
        "200":
          description: Accepted
  /v1/signal/inbox:
    get:
      summary: Pull queued signaling events (fallback mode)
      operationId: pullSignalInbox
      parameters:
        - name: clientType
          in: query
          required: true
          schema:
            type: string
        - name: clientId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /v1/signal/stream:
    get:
      summary: Subscribe signaling events via SSE (scaffold phase)
      operationId: streamSignalEvents
      parameters:
        - name: clientType
          in: query
          required: true
          schema:
            type: string
        - name: clientId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream established
x-websocket:
  endpoints:
    - path: /ws/desktop
      description: Desktop persistent control channel
    - path: /ws/mobile
      description: Mobile persistent control channel
  events:
    - signal.offer
    - signal.answer
    - signal.ice
    - relay.message
    - ack